chatbot_backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain 
openai 
pandas
tabulate
python-dotenv
openpyxl
streamlit
pandasai
pypdf
tiktoken
sentence-transformers
faiss-cpu
langchainhub
langchain-openai
shutil